Create an abstract class that is used to visualize hierarchical data. Treemap operator and Sunburst chart operator overrides this class.
